+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * Add `Iterators.partition` support
    ([#3212](https://github.com/JuliaData/DataFrames.jl/pull/3212))
+* Add `allunique` and allow transformations in `cols` argument of `describe`
+  and `nonunique` when working with `SubDataFrame`
+  ([3232](https://github.com/JuliaData/DataFrames.jl/pull/3232))
 
 # DataFrames.jl v1.4.4 Patch Release Notes